requirements.txt
aiofile==3.8.5
aiofiles==23.1.0
aiohttp==3.8.4
aiopath==0.6.11
aiosignal==1.3.1
anyio==3.6.2
async-timeout==4.0.2
asyncclick==8.1.3.4
attrs==23.1.0
caio==0.9.12
charset-normalizer==3.1.0
colorama==0.4.6
Cython==0.29.34
frozenlist==1.3.3
idna==3.4
markdown-it-py==2.2.0
mdurl==0.1.2
multidict==6.0.4
peewee==3.16.2
Pillow==9.5.0
Pygments==2.15.1
rich==13.3.5
selectolax==0.3.13
sniffio==1.3.0
tqdm==4.65.0
yarl==1.9.2